Rename: authModal -> authModalContainer로 변경
